Move deadchat and wolfchat to their own module
